requirements.txt
fastapi==0.135.1
google_cloud_aiplatform==1.141.0
pandas==3.0.1
protobuf==7.34.0
pydantic==2.12.5
python-dotenv==1.2.2
uvicorn==0.41.0
